refactor(kb): extract ingest source MCP facade (#12033)
